+++ src/main/java/com/bayou/managers/impl/UserManager.java
@@ -3,10 +3,11 @@ package com.bayou.managers.impl;
 import com.bayou.converters.LoginConverter;
 import com.bayou.converters.UserConverter;
 import com.bayou.domains.User;
+import com.bayou.exceptions.VerificationException;
 import com.bayou.managers.IManager;
 import com.bayou.ras.impl.UserResourceAccessor;
-import com.bayou.views.impl.LoginView;
 import com.bayou.views.impl.UserView;
+import com.bayou.views.impl.VerifyUserView;
 import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -29,24 +30,26 @@ public class UserManager implements IManager<UserView> {
     @Autowired
     LoginConverter loginConverter = new LoginConverter();
 
-    public LoginView login(LoginView loginView) throws NotFoundException {
+    public UserView login(VerifyUserView verifyUserView) throws NotFoundException, VerificationException {
         User returnedUser;
-        LoginView newLoginView;
 
-        if (!(loginView.getEmail() == null)) { //if email field is not null, get the user by email
-            returnedUser = ras.findByEmail(loginView.getEmail());
-        } else if (!(loginView.getAccountName() == null)) { //if account name is not null, get the user by account name
-            returnedUser = ras.findByAccountName(loginView.getAccountName());
+        if(verifyUserView.getEmail() != null) { //if email field is not null, get the user by email
+            returnedUser = ras.findByEmail(verifyUserView.getEmail());
+        } else if(verifyUserView.getAccountName() != null) { //if account name is not null, get the user by account name
+            returnedUser = ras.findByAccountName(verifyUserView.getAccountName());
         } else {
-            throw new NotFoundException("Values of email or account name not found" + "email: " + loginView.getEmail() + " account name: " + loginView.getAccountName());
+            throw new NotFoundException("Values of email or account name not found" + "email: " + verifyUserView.getEmail() + " account name: " + verifyUserView.getAccountName());
         }
         if (returnedUser == null) {
             throw new NotFoundException("The requested user does not exist in the database");
         } else {
-            newLoginView = loginConverter.convertToLoginView(returnedUser);
+            verifyUserView.setPasswordHash(returnedUser.getPasswordHash());
         }
-
-        return newLoginView;
+        if(verifyUserView.login()) {
+            return converter.convertToView(returnedUser);
+        } else {
+            throw new VerificationException();
+        }
     }
 
     public UserView get(Long id) throws NotFoundException {
